Analysis: Add a GenericCycleInfo analysis
Summary:
This analysis allows algorithms to capture all loops, including irreducible
loops, in a structured way. Based on Havlak (1997), Nesting of Reducible and
Irreducible Loops. See the definition of a "cycle" in the header file and
the IR test cases for more detail.

This patch also immediately adds (new and old-style) IR function passes.

Original implementation by Nicolai Hähnle <nicolai.haehnle@amd.com>
https://reviews.llvm.org/D83094

This change reintroduces the same implementation, but based on static
polymorphism that is readily available in the existing codebase.

Change-Id: I9c26fb1ce0b783e39790780039b7ad88596575c1